refactor: 2차 코멘트 반영 커밋
