+++ src/test/kotlin/racing/CarTest.kt
@@ -21,7 +21,8 @@ class CarTest : FunSpec({
             val carInstance = Car(5)
             // when, then
             for (i in 0..5) {
-                carInstance.currentLocation shouldBe i
+                carInstance.move()
+                carInstance.currentLocation shouldBe i + 1
             }
         }
     }
